requirements/base.txt
python-dotenv==1.2.1 
djangorestframework==3.16.1 
django-autoslug==1.9.9 
django-countries==8.2.0 
django-phonenumber-field==8.4.0 
phonenumbers==9.0.24 
drf-yasg==1.21.14 
pillow==12.1.1 
argon2-cffi==25.1.0 
djoser==2.3.3 
django-taggit==6.1.0 
django-filter==25.2 
cloudinary==1.44.1 
django-celery-email==3.0.0 
django-celery-beat==2.8.1 
redis==7.2.0 
celery==5.6.2 
flower==2.0.1